+++ src/desktopappScripts/watchElectron.ts
@@ -54,11 +54,11 @@ export async function watchElectron({
       
       entry: {
         main: path.join(cwd, 'src', app, 'main'),
-        libraryTarget: 'commonjs2',
       },
       
       output: {
         path: path.join(output, 'electron'),
+        libraryTarget: 'commonjs2',
       },
     },
     createWebpackWebappConfig({